pyproject.toml
[tool.poetry]
name = "multiviewae"
version = "0.0.0"
description = "A library for running multiview autoencoder models"
authors = ["analawryaguila <ana.lawryaguila@outlook.com>, Alejandra Jayme <alejandra.jayme@icloud.com>"]
readme = "README.md"
packages = [{include = "multiviewae"}]

license = "MIT"
documentation = "https://multiviewae.readthedocs.io/en/latest/"

[tool.setuptools.package-data]
"" = ["*.yaml"]

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^5.4.1"
pytorch-lightning = "1.5.10"
torch= "*"
torchvision = "*"
tensorboard = "*"
scipy = "*"
pandas = "*"
hydra-core = "*"
schema = "*"
numpy = "*"
matplotlib = "*"
ordereddict = "*"
pytest-cov = "*"

[build-system]
requires = ["poetry-core", "setuptools ~= 58.0"]
build-backend = "poetry.core.masonry.api"
